Refactor pipeline manager tests into separate files
